+++ software/mbc5.h
@@ -13,9 +13,9 @@
 #define MBC5_RUMBLE_BIT              0b100
 
 // Public functions
-uint16_t mbc5_memcpy_rom(uint8_t* dest, uint16_t rom_addr, uint16_t num);
+uint16_t mbc5_memcpy_rom(uint8_t* dest, uint32_t rom_addr, uint32_t num);
 // Note: This gets memory relative to RAM, not the cart. So 0x0 means start of RAM
-uint16_t mbc5_memcpy_ram(uint8_t* dest, uint16_t ram_addr, uint16_t num);
+uint16_t mbc5_memcpy_ram(uint8_t* dest, uint32_t ram_addr, uint32_t num);
 void mbc5_rom_dump(uint8_t *buf, uint16_t start_bank, uint16_t end_bank);
 void mbc5_ram_dump(uint8_t *buf, uint8_t start_bank, uint8_t end_bank);
